RavenDB-5244 V4.0 new client : Remove temp DatabaseCommands
